Fixed as Jenkins errors.

+++ sam/drivers/spi/spi.h
@@ -502,7 +502,7 @@ void spi_set_clock_phase(Spi *p_spi, uint32_t ul_pcs_ch, uint32_t ul_phase);
 void spi_configure_cs_behavior(Spi *p_spi, uint32_t ul_pcs_ch,
 		uint32_t ul_cs_behavior);
 void spi_set_bits_per_transfer(Spi *p_spi, uint32_t ul_pcs_ch, uint32_t ul_bits);
-int16_t spi_calc_baudrate_div(const uint32_t baudrate, uint32_t mck);
+int32_t spi_calc_baudrate_div(const uint32_t baudrate, uint32_t mck);
 int16_t spi_set_baudrate_div(Spi *p_spi, uint32_t ul_pcs_ch,
 		uint8_t uc_baudrate_divider);
 void spi_set_transfer_delay(Spi *p_spi, uint32_t ul_pcs_ch, uint8_t uc_dlybs,
